Change from .tokenUri("0") to .getDogTokenUris("0")
I noticed that in the fulfill random words test, the portion of the test that grabs the token URI, doesn't grab the Dog Token Uris with the correct method. I also changed it from inputting a number to a string to make it more like the rest of the code.